Address and Employee Service modifications
Check arguments of methods and pass real objects when needed.

+++ GSSS-Backend/src/main/java/ca/mcgill/ecse321/GSSS/service/AddressService.java
@@ -1,6 +1,7 @@
 package ca.mcgill.ecse321.GSSS.service;
 
 import java.util.List;
+import java.util.UUID;
 
 import javax.transaction.Transactional;
 
@@ -26,18 +27,32 @@ public class AddressService {
 	 * @param streetNumber Street number of the address
 	 * @param city City of the address
 	 * @param postalCode PostalCode of the address
-	 * @param id Id of the address
 	 * @return The new created Address
 	 */
 	@Transactional
-	public Address createAddress(String fullName, String streetName, Integer streetNumber, String city, String postalCode, String id) {
+	public Address createAddress(String fullName, String streetName, Integer streetNumber, String city, String postalCode) {
+		// Input validation
+	    String error = "";
+	    if(fullName == null || fullName.trim().length() == 0)
+	      error += "Address full name cannot be empty! ";
+	    if(streetName == null || streetName.trim().length() == 0)
+	      error += "Address street name cannot be empty! ";
+	    if(streetNumber == null)
+	      error += "Address street number cannot be empty! ";
+	    if(city == null || city.trim().length() == 0)
+	      error += "Address city cannot be empty! ";
+	    if(postalCode == null || postalCode.trim().length() == 0)
+		      error += "Address postal code cannot be empty! ";
+	    if(error.length() > 0)
+	      throw new IllegalArgumentException(error);
+		
 		Address address = new Address();
 		address.setFullName(fullName);
 		address.setStreetName(streetName);
 		address.setStreetNumber(streetNumber);
 		address.setCity(city);
 		address.setPostalCode(postalCode);
-		address.setId(id);
+		address.setId(UUID.randomUUID().toString());
 		addressRepository.save(address);
 		return address;
 	}
@@ -53,6 +68,10 @@ public class AddressService {
 	 */
 	@Transactional
 	public Address getAddress(String id) {
+		// Input validation
+	    if(id == null || id.trim().length() == 0)
+	      throw new IllegalArgumentException("Address id cannot be empty!");
+	    
 		return addressRepository.findAddressById(id);
 	}
 	
@@ -65,6 +84,11 @@ public class AddressService {
 	 */
 	@Transactional
 	public List<Address> getAddressByCity(String city) {
+		
+		// Input validation
+	    if(city == null || city.trim().length() == 0)
+	      throw new IllegalArgumentException("Address city cannot be empty!");
+	    
 		return addressRepository.findAddressesByCity(city);
 	}
 	
@@ -77,6 +101,10 @@ public class AddressService {
 	 */
 	@Transactional
 	public List<Address> getAddressByStreetName(String streetName) {
+		// Input validation
+	    if(streetName == null || streetName.trim().length() == 0)
+	      throw new IllegalArgumentException("Address street name cannot be empty!");
+	    
 		return addressRepository.findAddressesByStreetName(streetName);
 	}
 	
@@ -101,6 +129,10 @@ public class AddressService {
 	 */
 	@Transactional
 	public void deleteAddress(String id) {
+		// Input validation
+	    if(id == null || id.trim().length() == 0)
+	      throw new IllegalArgumentException("Address id cannot be empty!");
+	    
 		addressRepository.deleteById(id);
 	}
 	
@@ -120,7 +152,25 @@ public class AddressService {
 	 */
 	@Transactional
 	public Address modifyAddress(String fullName, String streetName, Integer streetNumber, String city, String postalCode, String id) {
-		Address address = addressRepository.findAddressById(id);
+		
+		// Input validation
+	    String error = "";
+	    if(fullName == null || fullName.trim().length() == 0)
+	      error += "Address full name cannot be empty! ";
+	    if(streetName == null || streetName.trim().length() == 0)
+	      error += "Address street name cannot be empty! ";
+	    if(streetNumber == null)
+	      error += "Address street number cannot be empty! ";
+	    if(city == null || city.trim().length() == 0)
+	      error += "Address city cannot be empty! ";
+	    if(postalCode == null || postalCode.trim().length() == 0)
+		      error += "Address postal code cannot be empty! ";
+	    if(id == null || id.trim().length() == 0)
+		      error += "Address id cannot be empty! ";
+	    if(error.length() > 0)
+	      throw new IllegalArgumentException(error);
+		
+		Address address = getAddress(id);
 		address.setFullName(fullName);
 		address.setStreetName(streetName);
 		address.setStreetNumber(streetNumber);
